refactor: fix all clippy warnings and clean dead code
- Remove 19 unused imports across 17 files
- Fix 13 unused/mut variable warnings (prefix with _)
- Replace 14 map_or(false, ...) with is_some_and(...)
- Replace 12 manual Default impls with #[derive(Default)]
- Fix manual_strip/manual_split_once -> idiomatic methods
- Remove redundant to_string() in format args
- Fix redundant closures, large enum variants (Box<Runtime>)
- Fix manual_div_ceil -> .div_ceil()
- Add #[allow(non_camel_case_types)] for GGUF spec names
- Update MSRV to 1.80 (LazyLock requirement)
- Zero clippy warnings, zero compile warnings, 250 tests pass

+++ src/inference/gguf.rs
@@ -419,7 +419,7 @@ impl GgufBackend for LlamaCppBackend {
 
         while token_count < request.max_tokens {
             // Sample next token
-            let new_token = sampler.sample(&ctx, (batch.n_tokens() - 1) as i32);
+            let new_token = sampler.sample(&ctx, batch.n_tokens() - 1);
             sampler.accept(new_token);
 
             // Check for end of sequence
@@ -542,7 +542,7 @@ impl GgufBackend for LlamaCppBackend {
 
         while token_count < request.max_tokens {
             // Sample next token
-            let new_token = sampler.sample(&ctx, (batch.n_tokens() - 1) as i32);
+            let new_token = sampler.sample(&ctx, batch.n_tokens() - 1);
             sampler.accept(new_token);
 
             // Check for end of sequence
